Increase Render Distance
- Increase base draw distance to 5000.
- Increase max draw distance to 15000.
- Add medium draw distance for certain levels.
- Remove cyclorama from depth buffer.
- Remove cyclorama requirement from gLevelClearColor.
- Change the type of gNumEnemyOfKind to int32_t.
- Raise MAX_SUPERTILE_ACTIVE_RANGE to 20.
- Change all values in gLevelSuperTileActiveRange to fit in the 8-20 range.
- Raise MAX_LIQUID_MESHES to 255.
Issue https://github.com/jorio/Bugdom/issues/24.

+++ src/Headers/game.h
@@ -199,7 +199,7 @@ extern	short						gMoney;
 extern	short						gNumBlueClovers;
 extern	short						gNumCollisions;
 extern	short						gNumEnemies;
-extern	short						gNumEnemyOfKind[];
+extern	int32_t						gNumEnemyOfKind[];
 extern	short						gNumGoldClovers;
 extern	short						gNumGreenClovers;
 extern	short						gNumLadyBugsThisArea;
